backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
email-validator==2.0.0
asyncio==3.4.3
httpx==0.23.3  # Changed from 0.24.1 to be compatible with solana
redis==4.6.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
solana==0.30.2
base58==2.1.1
cryptography==41.0.3
websockets==11.0.3
aiohttp==3.8.5
